refactor: move client-side logic to DisplayItemDetailsClient componen
relates #291
