+++ app/(dashboard)/item/[id]/page.tsx
@@ -1,13 +1,6 @@
 import newServerClient from '@/supabase/utils/newServerClient';
-
-//Components
-import Image from 'next/image';
-import ItemDetails from '@/components/ItemDetails';
-import PostageOptionDisplay from '@/components/PostageOptionDisplay';
-import BackButton from '@/components/buttons/BackButton';
 import { getProfile } from '@/supabase/models/getProfile';
-import NewConversationButton from '@/components/buttons/NewConversationButton';
-import ReserveButton from '@/components/buttons/ReserveButton';
+import DisplayItemDetailsClient from '@/components/DisplayItemDetailsClient';
 
 const DisplayItemDetails = async ({ params }: { params: { id: string } }) => {
   const supabase = newServerClient();
@@ -26,7 +19,7 @@ const DisplayItemDetails = async ({ params }: { params: { id: string } }) => {
       throw new Error('Error fetching data');
     } else {
       const donorEmail = item.profiles.email;
-      const donerId: string = item.profiles?.id;
+      const donorId: string = item.profiles?.id;
       const title = item.item_name;
       if (
         !user ||
@@ -38,54 +31,14 @@ const DisplayItemDetails = async ({ params }: { params: { id: string } }) => {
       }
 
       return (
-        <>
-          <BackButton />
-          <div className='mb-10 mt-2 flex flex-col items-center gap-14'>
-            <div className='relative h-52 w-72 md:h-72 md:w-96'>
-              <Image
-                src={`${item.imageSrc}`}
-                alt={`${item.item_name}`}
-                layout='fill'
-                objectFit='cover'
-                className='shadow-md'
-              />
-            </div>
-            <PostageOptionDisplay
-              collectible={item.collectible}
-              postable={item.postable}
-              postage_covered={item.postage_covered}
-            />
-            <div className='min-h-40 w-full bg-secondaryGray p-10 md:w-1/2 md:rounded-lg'>
-              <div className='flex flex-row justify-between'>
-                <h2 className='place-self-center text-xl italic'>
-                  {item.item_name}
-                </h2>
-                {item.reserved && <p className='reserved'>Reserved</p>}
-              </div>
-              <h3 className='pt-3 font-light'>Description:</h3>
-              <p className='pt-2 text-center'>{item.item_description}</p>
-            </div>
-            <ItemDetails
-              condition={item.condition}
-              donated_by={item.profiles.username}
-              postcode={item.postcode}
-              fontSize='text-lg'
-            />
-
-            {canMessage && user?.id && item.id && (
-              <div className='flex flex-row gap-4 pb-10'>
-                <NewConversationButton
-                  userId={user?.id}
-                  donorId={donerId}
-                  donorEmail={donorEmail}
-                  title={title}
-                  item_id={item.id}
-                />
-                <ReserveButton itemId={item.id} userId={user.id} />
-              </div>
-            )}
-          </div>
-        </>
+        <DisplayItemDetailsClient
+          item={item}
+          user={user!}
+          canMessage={canMessage}
+          donorId={donorId}
+          donorEmail={donorEmail}
+          title={title}
+        />
       );
     }
   } catch (error) {
